requirements.txt
flask
openstacksdk
requests

